% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-info.R
\name{asn_info}
\alias{asn_info}
\title{Search for an autonomous system}
\usage{
asn_info(asn, api_key = networksdb_api_key())
}
\arguments{
\item{asn}{the autonomous system number}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Search for an autonomous system
}
