% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn-networks.R
\name{asn_nets}
\alias{asn_nets}
\title{Search for the networks announced by an autonomous system}
\usage{
asn_nets(asn, page = 1, api_key = networksdb_api_key())
}
\arguments{
\item{asn}{the autonomous system number}

\item{page}{results pagination}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Search for the networks announced by an autonomous system
}
