% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains-in-network.R
\name{domains_in_network}
\alias{domains_in_network}
\title{Perform a "mass" reverse DNS lookup to find all the domain names pointing to any IPv4 or IPv6 address in the given network}
\usage{
domains_in_network(
  ip_start,
  ip_end,
  cidr,
  page = 1,
  api_key = networksdb_api_key()
)
}
\arguments{
\item{ip_start, ip_end, cidr}{Either use \code{ip_start} and \code{ip_end} OR \code{cidr};}

\item{page}{results pagination}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Perform a "mass" reverse DNS lookup to find all the domain names pointing to any IPv4 or IPv6 address in the given network
}
