% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains-on-ip.R
\name{domains_on_ip}
\alias{domains_on_ip}
\title{Perform a reverse DNS lookup to find all the domain names pointing to the given IPv4 or IPv6 address}
\usage{
domains_on_ip(ip, page = 1, api_key = networksdb_api_key())
}
\arguments{
\item{ip}{address to lookup}

\item{page}{pagination}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Perform a reverse DNS lookup to find all the domain names pointing to the given IPv4 or IPv6 address
}
