% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-geo.R
\name{ip_geo}
\alias{ip_geo}
\title{Request geolocation information for a given IPv4 or IPv6 Address.}
\usage{
ip_geo(ip, api_key = networksdb_api_key())
}
\arguments{
\item{ip}{IPv4 or IPv6 address}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Request geolocation information for a given IPv4 or IPv6 Address.
}
