% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-info.R
\name{ip_info}
\alias{ip_info}
\title{Search for an IPv4 or IPv6 address}
\usage{
ip_info(ip, api_key = networksdb_api_key())
}
\arguments{
\item{ip}{IPv4 or IPv6 address}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Search for an IPv4 or IPv6 address
}
