% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-for-domain.R
\name{ips_for_domain}
\alias{ips_for_domain}
\title{Perform a forward DNS lookup to find all the IPv4 and IPv6 addresses pointed to by a given domain name}
\usage{
ips_for_domain(domain, api_key = networksdb_api_key())
}
\arguments{
\item{domain}{domain to lookup}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Perform a forward DNS lookup to find all the IPv4 and IPv6 addresses pointed to by a given domain name
}
