% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksdb-package.R
\docType{package}
\name{networksdb}
\alias{networksdb}
\alias{networksdb-package}
\title{Query Organization, IP, DNS, and Network Info From NetworksDB}
\description{
NetworksDB \url{https://networksdb.io} contains information about the public IPv4 and
IPv6 addresses, networks and domains owned by companies and organisations across the world
along with city-level IP geolocation data and autonomous system information. Tools are provided
to query the NetworksDB API.
}
\references{
\url{https://networksdb.io/api/docs}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/networksdb}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/networksdb/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
