% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksdb-pat.R
\name{networksdb_api_key}
\alias{networksdb_api_key}
\title{Get or set NetworksDB Personal Access Token}
\usage{
networksdb_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new NetworksDB API Key}
}
\value{
atomic character vector containing the NetworksDB API key
}
\description{
Get or set NetworksDB Personal Access Token
}
\references{
\url{https://networksdb.io/api/order/free}
}
