% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-info.R
\name{org_info}
\alias{org_info}
\title{Retreive information about an organisation.}
\usage{
org_info(id, api_key = networksdb_api_key())
}
\arguments{
\item{id}{id of an organisation (i.e. returned by \code{\link[=org_search]{org_search()}})}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Retreive information about an organisation.
}
