% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-networks.R
\name{org_networks}
\alias{org_networks}
\title{Search for the public networks owned by an organisation.}
\usage{
org_networks(id, page = 1, api_key = networksdb_api_key())
}
\arguments{
\item{id}{id of an organisation (i.e. returned by \code{\link[=org_search]{org_search()}})}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\description{
Search for the public networks owned by an organisation.
}
