% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-search.R
\name{org_search}
\alias{org_search}
\title{Search for organisations matching a search term.}
\usage{
org_search(query, country = NULL, page = 1, api_key = networksdb_api_key())
}
\arguments{
\item{query}{search terms}

\item{country}{if not \code{NULL} (the default) a single ISO2C country code}

\item{page}{page number (the query result set may have multiple pages; the current page number
of the result is a slot in the return value list.)}

\item{api_key}{see \code{\link[=networksdb_api_key]{networksdb_api_key()}}}
}
\value{
\code{list} with \code{total} (# of results for the search), \code{page} (the current page number), and
\code{results} which is a data frame of search results. if \code{nrow()} of this data frame is >
\code{total} then there are more pages available.
}
\description{
Search for organisations matching a search term.
}
