% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksdb-package.R
\docType{package}
\name{networksdb}
\alias{networksdb}
\alias{networksdb-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/networksdb}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/networksdb/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
